pyproject.toml
[tool.poetry]
name = "elts"
version = "0.1.0"
description = ""
authors = ["Tom Bischopink <s87622@informatik.htw-dresden.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "3.9.3"
aiosignal = "1.3.1"
aiosmtplib = "2.0.2"
alembic = "1.13.1"
annotated-types = "0.6.0"
anyio = "4.3.0"
async-timeout = "4.0.3"
attrs = "23.2.0"
autopep8 = "2.0.4"
bcrypt = "4.1.2"
blinker = "1.7.0"
certifi = "2024.2.2"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
cryptography = "42.0.5"
dataclasses-json = "0.6.4"
distro = "1.9.0"
dnspython = "2.6.1"
ecdsa = "0.18.0"
email-validator = "2.1.1"
fastapi = "0.110.0"
fastapi-mail = "1.4.1"
frozenlist = "1.4.1"
h11 = "0.14.0"
httpcore = "1.0.4"
httpx = "0.27.0"
idna = "3.6"
iniconfig = "2.0.0"
jinja2 = "3.1.3"
jose = "1.0.0"
jsonpatch = "1.33"
jsonpointer = "2.4"
jwt = "1.3.1"
langchain = "0.1.10"
langchain-community = "0.0.25"
langchain-core = "0.1.28"
langchain-text-splitters = "0.0.1"
langsmith = "0.1.16"
mako = "1.3.2"
markupsafe = "2.1.5"
marshmallow = "3.21.1"
marshmallow-enum = "1.5.1"
multidict = "6.0.5"
mypy-extensions = "1.0.0"
numexpr = "2.9.0"
numpy = "1.26.4"
openai = "1.13.3"
openapi-schema-pydantic = "1.2.4"
orjson = "3.9.15"
packaging = "23.2"
passlib = "1.7.4"
pluggy = "1.4.0"
psycopg2-binary = "2.9.9"
pyasn1 = "0.5.1"
pycodestyle = "2.11.1"
pycparser = "2.21"
pycrypto = "2.6.1"
pydantic = "2.6.3"
pydantic-settings = "2.2.1"
pydantic-core = "2.16.3"
pymongo = "4.6.2"
pypdf = "4.1.0"
pytest = "8.0.2"
python-dotenv = "1.0.1"
python-jose = "3.3.0"
python-multipart = "0.0.9"
pyyaml = "6.0.1"
regex = "2023.12.25"
requests = "2.31.0"
rsa = "4.9"
six = "1.16.0"
sniffio = "1.3.1"
sqlalchemy = "2.0.28"
starlette = "0.36.3"
tenacity = "8.2.3"
tiktoken = "0.6.0"
tomli = "2.0.1"
tqdm = "4.66.2"
typing-inspect = "0.9.0"
typing-extensions = "4.10.0"
urllib3 = "2.2.1"
uvicorn = "0.27.1"
yarl = "1.9.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
